updated YamlReader. fix RemoveById

+++ src/main/java/commands/RemoveById.java
@@ -12,7 +12,7 @@ public class RemoveById extends BaseCommand {
     @Override
     public void execute(DataBase obj) throws IOException {
         try{
-            Long a = Long.parseLong(super.getParameter());
+            Integer a = Integer.parseInt(super.getParameter());
             var array = obj.getCollection();
             Iterator<Person> it = array.iterator();
             while(it.hasNext()){
